fix: applied suggested changes
