+++ packages/market-service/src/index.ts
@@ -1,2 +1,3 @@
 export * from './market-service-manager'
 export * from './fiat-market-service-manager'
+export * from './fiat-market-service-types'
